refactor: devotion banner bundle
