+++ src/assets/js/devotionBannerModel.js
